backend/requirements.txt
django==5.0.3
djangorestframework==3.15.1
psycopg2-binary==2.9.9
django-redis==5.4.0
gunicorn==21.2.0
pytest==8.3.3
pytest-django==4.8.0
requests==2.32.3
